code refactoring, fixing compiler warnings -Wverything

+++ src/vita/system/fileio.c
@@ -19,7 +19,7 @@ str_t *file_read(const char *const z_filename) {
     }
 
     // get file size
-    const size_t fsize = path_get_file_size(z_filename);
+    const int64_t fsize = path_get_file_size(z_filename);
     if(fsize < 0) {
         DEBUG_ASSERT(fsize >= 0, "Unable to fetch file size data!");
